# Refactor crm report