--- a/data-manifest
+++ b/data-manifest
@@ -70,6 +70,8 @@ test/features/bootstrap_init_join_remove.feature
 test/features/bootstrap_options.feature
 test/features/environment.py
 test/features/geo_setup.feature
+test/features/hb_report_bugs.feature
+test/features/hb_report_normal.feature
 test/features/qdevice_options.feature
 test/features/qdevice_setup_remove.feature
 test/features/qdevice_usercase.feature