--- a/data-manifest
+++ b/data-manifest
@@ -182,10 +182,12 @@ test/unittests/test.conf
 test/unittests/test_corosync.py
 test/unittests/test_gv.py
 test/unittests/test_handles.py
+test/unittests/test_hb_report_collect.py
+test/unittests/test_hb_report_core.py
+test/unittests/test_hb_report_utils.py
 test/unittests/test_objset.py
 test/unittests/test_parallax.py
 test/unittests/test_parse.py
-test/unittests/test_report.py
 test/unittests/test_scripts.py
 test/unittests/test_time.py
 test/unittests/test_utils.py