@zzhou1 2020-09-29
You may want to add bsc#1176441 into the bug list for tracking too.